+++ src/gtirb-decoder/target/RawX86Loader.h
@@ -1,6 +1,6 @@
-//===- PeReader.h       -----------------------------------------*- C++ -*-===//
+//===- RawX86Loader.h -------------------------------------------*- C++ -*-===//
 //
-//  Copyright (C) 2020 GrammaTech, Inc.
+//  Copyright (C) 2021 GrammaTech, Inc.
 //
 //  This code is licensed under the GNU Affero General Public License
 //  as published by the Free Software Foundation, either version 3 of
@@ -20,30 +20,22 @@
 //  endorsement should be inferred.
 //
 //===----------------------------------------------------------------------===//
-#ifndef PE_GTIRB_BUILDER_H_
-#define PE_GTIRB_BUILDER_H_
+#ifndef SRC_GTIRB_DECODER_TARGET_RAWX86LOADER_H_
+#define SRC_GTIRB_DECODER_TARGET_RAWX86LOADER_H_
 
-#include "./GtirbBuilder.h"
+#include "../CompositeLoader.h"
+#include "../core/DataLoader.h"
+#include "../format/RawLoader.h"
 
-class PeReader : public GtirbBuilder
+CompositeLoader RawX86Loader()
 {
-public:
-    PeReader(std::string Path, std::shared_ptr<LIEF::Binary> Binary);
+    CompositeLoader Loader("souffle_disasm_x86_86");
+    Loader.add(ModuleLoader);
+    Loader.add(SectionLoader);
+    Loader.add<X86Loader>();
+    Loader.add<DataLoader>(DataLoader::Pointer::DWORD);
+    Loader.add(RawEntryLoader);
+    return Loader;
+}
 
-    gtirb::ErrorOr<GTIRB> build() override;
-
-protected:
-    std::shared_ptr<LIEF::PE::Binary> Pe;
-
-    void initModule() override;
-    void buildSections() override;
-    void buildSymbols() override;
-    void addEntryBlock() override;
-    void addAuxData() override;
-
-    std::vector<PeResource> resources();
-    std::vector<ImportEntry> importEntries();
-    std::vector<ExportEntry> exportEntries();
-};
-
-#endif // PE_GTIRB_BUILDER_H_
+#endif // SRC_GTIRB_DECODER_TARGET_RAWX86LOADER_H_