+++ tests/misc_test.py
@@ -1,6 +1,14 @@
+import os
 import platform
 import unittest
-from disassemble_reassemble_check import compile, cd, disassemble
+import subprocess
+from disassemble_reassemble_check import (
+    compile,
+    cd,
+    disassemble,
+    reassemble,
+    test,
+)
 from pathlib import Path
 import gtirb
 
@@ -98,6 +106,55 @@ class AuxDataTests(unittest.TestCase):
                     break
             assert found
 
+    @unittest.skipUnless(
+        platform.system() == "Linux", "This test is linux only."
+    )
+    def test_souffle_relations(self):
+        """Test `--with-souffle-relations' equivalence to `--debug-dir'."""
+
+        with cd(ex_dir / "ex1"):
+            # build
+            self.assertTrue(compile("gcc", "g++", "-O0", []))
+
+            # disassemble
+            if not os.path.exists("dbg"):
+                os.mkdir("dbg")
+            self.assertTrue(
+                disassemble(
+                    "ex",
+                    False,
+                    False,
+                    False,
+                    format="--ir",
+                    extension="gtirb",
+                    extra_args=[
+                        "-F",
+                        "--with-souffle-relations",
+                        "--debug-dir",
+                        "dbg",
+                    ],
+                )
+            )
+
+            # load the gtirb
+            ir = gtirb.IR.load_protobuf("ex.gtirb")
+            m = ir.modules[0]
+
+            # dump relations to directory
+            if not os.path.exists("aux"):
+                os.mkdir("aux")
+            for table, ext in [
+                ("souffleFacts", "facts"),
+                ("souffleOutputs", "csv"),
+            ]:
+                for name, relation in m.aux_data[table].data.items():
+                    _, csv = relation
+                    with open(f"aux/{name}.{ext}", "w") as out:
+                        out.write(csv)
+
+            # compare the relations directories
+            subprocess.check_call(["diff", "dbg", "aux"])
+
 
 class MovedLabelTests(unittest.TestCase):
     @unittest.skipUnless(
@@ -142,6 +199,45 @@ class MovedLabelTests(unittest.TestCase):
             assert isinstance(se1, gtirb.SymAddrConst)
             self.assertEqual(se1.symbol.name, "point.2")
             self.assertEqual(se1.offset, 22)
+
+
+class RawGtirbTests(unittest.TestCase):
+    @unittest.skipUnless(
+        platform.system() == "Linux", "This test is linux only."
+    )
+    def test_read_gtirb(self):
+
+        binary = "ex"
+        with cd(ex_dir / "ex1"):
+            self.assertTrue(compile("gcc", "g++", "-O0", []))
+
+            # Output GTIRB file without disassembling.
+            self.assertTrue(
+                disassemble(
+                    binary,
+                    False,
+                    False,
+                    False,
+                    format="--ir",
+                    extension="gtirb",
+                    extra_args=["--no-analysis"],
+                )
+            )
+
+            # Disassemble GTIRB input file.
+            self.assertTrue(
+                disassemble(
+                    "ex.gtirb",
+                    False,
+                    False,
+                    False,
+                    format="--asm",
+                    extension="s",
+                )
+            )
+
+            self.assertTrue(reassemble("gcc", "ex.gtirb", extra_flags=[]))
+            self.assertTrue(test())
 
 
 if __name__ == "__main__":